% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsdbImport.R
\name{bsdb_import}
\alias{bsdb_import}
\title{Obtain published microbial signatures from bugsigdb.org database
using bugsigdb. The obtained microbes is filtered based on the user inputs.}
\usage{
bsdb_import(
  location = "United States of America",
  condition = "obesity",
  body_site = "feces"
)
}
\arguments{
\item{location}{The regions that the samples from participants
were taken from(e.g United States of America)}

\item{condition}{Area/Field of study(e.g obesity)}

\item{body_site}{Body site that the sample was taken from (e.g fecal\link{feces})}
}
\value{
a \code{\link{data.frame}}.
}
\description{
Obtain published microbial signatures from bugsigdb.org database
using bugsigdb. The obtained microbes is filtered based on the user inputs.
}
\references{
BugSigDB: \url{https://bugsigdb.org}
}
