% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageTemplate.R
\docType{package}
\name{packageTemplate}
\alias{packageTemplate}
\title{ihmeuw-demographics R package template}
\description{
ihmeuw-demographics R package template that is used when
creating new R packages for our team.
}
