% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/placeholder.R
\name{placeholder}
\alias{placeholder}
\title{Place holder function that does nothing}
\usage{
placeholder(x)
}
\arguments{
\item{x}{default input}
}
\value{
returns the input \code{x} unmodified
}
\description{
Place holder function that does nothing
}
\examples{
placeholder(0)

}
